+++ report/presentation/src/main/kotlin/me/proton/core/report/presentation/entity/internal_entities.kt
@@ -30,9 +30,7 @@ internal sealed class BugReportFormState {
 
 internal data class ReportFormData(
     val subject: String,
-    val description: String,
-    val country: String? = null,
-    val isp: String? = null
+    val description: String
 )
 
 internal enum class ExitSignal {