+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Current plan used space color indicator.
 - Default plan max space to take into account RewardsMaxSpace.
 - Use user currency in the current plan subscription screen for free plan.
+- Improve displaying error messages in Bug Report form
 
 ## [7.1.11]